# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/products": {
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filtrar por slug de categoría",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      }
    },
    "/api/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      }
    },
    "/api/products/admin/all": {
      "get": {
        "operationId": "ProductsController_findAllAdmin",
        "parameters": [
          {
            "name": "includeInactive",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      }
    },
    "/api/products/admin/{id}": {
      "get": {
        "operationId": "ProductsController_findOneAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      }
    },
    "/api/products/{id}/restore": {
      "patch": {
        "operationId": "ProductsController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      }
    },
...
    "/api/reviews/product/{productId}": {
      "get": {
        "operationId": "ReviewsController_findByProduct",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "query",
            "description": "ID del producto",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/reviews": {
      "post": {
        "operationId": "ReviewsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/reviews/{id}": {
      "delete": {
        "operationId": "ReviewsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/reviews/admin/all": {
      "get": {
        "operationId": "ReviewsController_findAllAdmin",
        "parameters": [
          {
            "name": "includeInactive",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/reviews/{id}/restore": {
      "patch": {
        "operationId": "ReviewsController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProductCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Remeras"
          },
          "slug": {
            "type": "string",
            "example": "remeras"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "AttributeType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Talle"
          },
          "slug": {
            "type": "string",
            "example": "talle"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "AttributeValue": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "value": {
            "type": "string",
            "example": "M"
          },
          "attribute_types": {
            "$ref": "#/components/schemas/AttributeType"
          }
        },
        "required": [
          "id",
          "value",
          "attribute_types"
        ]
      },
      "VariantAttribute": {
        "type": "object",
        "properties": {
          "attribute_values": {
            "$ref": "#/components/schemas/AttributeValue"
          }
        },
        "required": [
          "attribute_values"
        ]
      },
      "ProductVariant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "sku": {
            "type": "object",
            "example": "SKU-001"
          },
          "price_modifier": {
            "type": "object",
            "example": 0
          },
          "stock": {
            "type": "number",
            "example": 10
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "variant_attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantAttribute"
            }
          }
        },
        "required": [
          "id",
          "sku",
          "price_modifier",
          "stock",
          "is_active",
          "variant_attributes"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Remera Técnica Elite"
          },
          "price": {
            "type": "number",
            "example": 25.99
          },
          "image": {
            "type": "string",
            "example": "https://example.com/image.png",
            "nullable": true
          },
          "description": {
            "type": "string",
            "example": "Descripción del producto",
            "nullable": true
          },
          "category_id": {
            "type": "number",
            "example": 1
          },
          "created_at": {
            "type": "string",
            "example": "2026-04-10T00:00:00.000Z",
            "nullable": true
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "categories": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductCategory"
              }
            ]
          },
          "product_variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariant"
            }
          }
        },
        "required": [
          "id",
          "name",
          "price",
          "image",
          "description",
          "category_id",
          "created_at",
          "is_active",
          "categories",
          "product_variants"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {}
      },
...
      "ReviewUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "4539152d-031b-41c4-8ef8-c419e8b5f84a"
          },
          "name": {
            "type": "string",
            "example": "Juan Pérez",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "Review": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "product_id": {
            "type": "number",
            "example": 1
          },
          "user_id": {
            "type": "string",
            "example": "4539152d-031b-41c4-8ef8-c419e8b5f84a"
          },
          "rating": {
            "type": "number",
            "example": 5,
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string",
            "example": "Excelente producto",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "example": "2026-04-10T00:00:00.000Z",
            "nullable": true
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "users": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewUser"
              }
            ]
          }
        },
        "required": [
          "id",
          "product_id",
          "user_id",
          "rating",
          "comment",
          "created_at",
          "is_active",
          "users"
        ]
      },
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "number",
            "example": 1
          },
          "rating": {
            "type": "number",
            "example": 5,
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "object",
            "example": "Excelente producto"
          }
        },
        "required": [
          "product_id",
          "rating"
        ]
      },
...
    }
  }
}